Cleaned up a few more warnings.
git-svn-id: svn://svn.code.sf.net/p/stella/code/trunk@3292 8b62c5a3-ac7e-4cc8-8f21-d9a121418aba

+++ src/emucore/TIA.hxx
@@ -217,7 +217,7 @@ class TIA : public Device
       based on how many frames of out the total count are PAL frames.
     */
     bool isPAL() const
-      { return float(myPALFrameCounter) / myFrameCounter >= (25.0/60.0); }
+      { return double(myPALFrameCounter) / myFrameCounter >= (25.0/60.0); }
 
     /**
       Answers the current color clock we've gotten to on this scanline.
